# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: eSIM reseller platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /family:
    get:
      operationId: listFamilyMembers
      tags: [family]
      summary: List family members
      responses:
        "200":
          description: Family member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyMember"
    post:
      operationId: createFamilyMember
      tags: [family]
      summary: Create a family member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyMemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyMember"
  /family/{memberId}:
    patch:
      operationId: updateFamilyMember
      tags: [family]
      summary: Update a family member
      parameters:
        - in: path
          name: memberId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyMemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyMember"
    delete:
      operationId: deleteFamilyMember
      tags: [family]
      summary: Delete a family member
      parameters:
        - in: path
          name: memberId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericSuccess"
...
  /referral/apply:
    post:
      operationId: applyReferral
      tags: [referral]
      summary: Apply a referral code during registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralApplyInput"
      responses:
        "200":
          description: Referral applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericSuccess"
  /redeem:
    post:
      operationId: redeemCode
      tags: [redeem]
      summary: Redeem a gift code for wallet credit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemInput"
      responses:
        "200":
          description: Code redeemed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /redeem/generate:
    post:
      operationId: generateRedeemCode
      tags: [redeem]
      summary: Generate a gift redeem code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCodeInput"
      responses:
        "201":
          description: Code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedeemCode"
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List support tickets
      responses:
        "200":
          description: Ticket list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{ticketId}/reply:
    post:
      operationId: replyTicket
      tags: [tickets]
      summary: Reply to a support ticket
      parameters:
        - in: path
          name: ticketId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketReplyInput"
      responses:
        "200":
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
...
components:
  schemas:
...
    Wallet:
      type: object
      required: [balanceUsd, level, levelName, totalSpentUsd, discountPercent]
      properties:
        balanceUsd:
          type: number
        level:
          type: string
          enum: [bronze, silver, gold, platinum]
        levelName:
          type: string
        totalSpentUsd:
          type: number
        discountPercent:
          type: integer
        nextLevelSpendRequired:
          type: ["number", "null"]
        levelupBonusUsd:
          type: number
...
    FamilyMember:
      type: object
      required: [id, name, age, avatar, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        age:
          type: integer
        avatar:
          type: string
        activeEsim:
          $ref: "#/components/schemas/ActiveEsimInfo"
        createdAt:
          type: string
    ActiveEsimInfo:
      type: object
      properties:
        orderNo:
          type: string
        locationName:
          type: string
        dataGb:
          type: number
        expiredTime:
          type: string
        status:
          type: string
    FamilyMemberInput:
      type: object
      required: [name, age, avatar]
      properties:
        name:
          type: string
        age:
          type: integer
        avatar:
          type: string
    FamilyMemberUpdate:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        avatar:
          type: string
...
    ReferralApplyInput:
      type: object
      required: [referralCode]
      properties:
        referralCode:
          type: string
    RedeemInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
    GenerateCodeInput:
      type: object
      required: [amountUsd]
      properties:
        amountUsd:
          type: number
    RedeemCode:
      type: object
      required: [code, amountUsd, status, createdAt]
      properties:
        code:
          type: string
        amountUsd:
          type: number
        status:
          type: string
          enum: [active, used]
        usedBy:
          type: ["string", "null"]
        usedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    Ticket:
      type: object
      required: [id, subject, message, status, createdAt]
      properties:
        id:
          type: string
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [pending, resolved]
        replies:
          type: array
          items:
            type: object
            required: [message, isAdmin, createdAt]
            properties:
              message:
                type: string
              isAdmin:
                type: boolean
              createdAt:
                type: string
        createdAt:
          type: string
    TicketInput:
      type: object
      required: [subject, message]
      properties:
        subject:
          type: string
        message:
          type: string
    TicketReplyInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    GenericSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...